Move DataView to components/index/cards/charts
