+++ components/index/cards/ConfirmedCasesDetailsCard.vue
@@ -45,7 +45,7 @@
 import dayjs from 'dayjs'
 import Data from '@/data/data.json'
 import formatConfirmedCases from '@/utils/formatConfirmedCases'
-import DataView from '@/components/DataView.vue'
+import DataView from '@/components/index/cards/charts/DataView.vue'
 import ConfirmedCasesDetailsTable from '@/components/index/cards/charts/ConfirmedCasesDetailsTable.vue'
 
 export default {
