+++ components/index/cards/charts/ConfirmedCasesByMunicipalitiesTable.vue
@@ -68,7 +68,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import DataView from '@/components/DataView.vue'
+import DataView from '@/components/index/cards/charts/DataView.vue'
 import DataViewBasicInfoPanel from '@/components/index/cards/charts/DataViewBasicInfoPanel.vue'
 
 export default Vue.extend({
