+++ components/index/cards/charts/DataTable.vue
@@ -113,7 +113,7 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import DataView from '@/components/DataView.vue'
+import DataView from '@/components/index/cards/charts/DataView.vue'
 import DataViewBasicInfoPanel from '@/components/index/cards/charts/DataViewBasicInfoPanel.vue'
 import OpenDataLink from '@/components/OpenDataLink.vue'
