+++ components/index/cards/charts/MetroBarChart.vue
@@ -69,7 +69,7 @@ import Vue from 'vue'
 import { TranslateResult } from 'vue-i18n'
 import { ChartOptions, ChartData, Chart } from 'chart.js'
 import { ThisTypedComponentOptionsWithRecordProps } from 'vue/types/options'
-import DataView from '@/components/DataView.vue'
+import DataView from '@/components/index/cards/charts/DataView.vue'
 import { getGraphSeriesStyle } from '@/utils/colors'
 import ExternalLink from '@/components/ExternalLink.vue'
 import { DisplayData } from '@/plugins/vue-chart'
