+++ components/index/cards/charts/TimeStackedBarChart.vue
@@ -101,7 +101,7 @@ import { ThisTypedComponentOptionsWithRecordProps } from 'vue/types/options'
 import { TranslateResult } from 'vue-i18n'
 import { Chart } from 'chart.js'
 import dayjs from 'dayjs'
-import DataView from '@/components/DataView.vue'
+import DataView from '@/components/index/cards/charts/DataView.vue'
 import DataSelector from '@/components/index/cards/charts/DataSelector.vue'
 import DataViewBasicInfoPanel from '@/components/index/cards/charts/DataViewBasicInfoPanel.vue'
 import { DisplayData, yAxesBgPlugin, scrollPlugin } from '@/plugins/vue-chart'
